FEAT: further improvement for class implementation
 - reworked a bit the game manager

 - now i can save the avatar, class and name which are important in character creation

+++ core/gui/images/images.cpp
@@ -41,5 +41,5 @@ CImage::~CImage() { };
 
 void CImage::Draw()
 {
-    _image.Draw(raylib::Rectangle(0, 0, _image.GetWidth(), _image.GetHeight()) ,_rectangle);
+    _image.Draw(raylib::Rectangle(0, 0, static_cast<float>(_image.GetWidth()), static_cast<float>(_image.GetHeight())) ,_rectangle);
 };
